feat(multitenancy): transfer older rabbitmq messages to new queues (#3505)
