+++ openaev-model/src/main/java/io/openaev/service/RabbitmqService.java
@@ -1,6 +1,5 @@
 package io.openaev.service;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.Connection;
@@ -11,15 +10,9 @@ import io.openaev.driver.RabbitmqDriver;
 import io.openaev.service.queue.BatchQueueService;
 import io.openaev.service.queue.QueueExecution;
 import io.openaev.service.queue.Queueable;
+import io.openaev.service.rabbitmq.RabbitmqManagementClient;
 import java.io.IOException;
-import java.net.URI;
-import java.net.URLEncoder;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.Base64;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -69,8 +62,7 @@ public class RabbitmqService {
   private final RabbitmqConfig rabbitmqConfig;
   private final ConnectionFactory connectionFactory;
   private final RabbitmqDriver rabbitmqDriver;
-  private final HttpClient httpClient =
-      HttpClient.newBuilder().connectTimeout(java.time.Duration.ofSeconds(30)).build();
+  private final RabbitmqManagementClient managementClient;
 
   // -- CONFIGURATION --
 
@@ -477,8 +469,6 @@ public class RabbitmqService {
 
   // -- MANAGEMENT API --
 
-  private static final ObjectMapper MAPPER = new ObjectMapper();
-
   /**
    * Lists all queue names from the RabbitMQ management API that start with the given prefix.
    *
@@ -487,7 +477,7 @@ public class RabbitmqService {
    * @throws IllegalStateException if the management API is unreachable or returns an error
    */
   public List<String> listQueueNamesWithPrefix(String prefix) {
-    return listNamesFromManagementApi("queues", prefix);
+    return managementClient.listQueueNamesWithPrefix(prefix);
   }
 
   /**
@@ -498,66 +488,7 @@ public class RabbitmqService {
    * @throws IllegalStateException if the management API is unreachable or returns an error
    */
   public List<String> listExchangeNamesWithPrefix(String prefix) {
-    return listNamesFromManagementApi("exchanges", prefix);
-  }
-
-  /**
-   * Queries the RabbitMQ management HTTP API for a list of resources (queues or exchanges) and
-   * returns names matching the given prefix.
-   *
-   * @throws IllegalStateException if the API is unreachable or returns a non-200 status
-   */
-  private List<String> listNamesFromManagementApi(String resource, String prefix) {
-    List<String> result = new ArrayList<>();
-    try {
-      String scheme = rabbitmqConfig.isSsl() ? "https" : "http";
-      String encodedVhost = URLEncoder.encode(rabbitmqConfig.getVhost(), StandardCharsets.UTF_8);
-      URI uri =
-          URI.create(
-              scheme
-                  + "://"
-                  + rabbitmqConfig.getHostname()
-                  + ":"
-                  + rabbitmqConfig.getManagementPort()
-                  + "/api/"
-                  + resource
-                  + "/"
-                  + encodedVhost
-                  + "?columns=name");
-
-      String credentials = rabbitmqConfig.getUser() + ":" + rabbitmqConfig.getPass();
-      String authHeader =
-          "Basic "
-              + Base64.getEncoder().encodeToString(credentials.getBytes(StandardCharsets.UTF_8));
-
-      HttpRequest request =
-          HttpRequest.newBuilder().uri(uri).header("Authorization", authHeader).GET().build();
-
-      HttpResponse<String> response =
-          httpClient.send(request, HttpResponse.BodyHandlers.ofString());
-
-      if (response.statusCode() == 200) {
-        JsonNode root = MAPPER.readTree(response.body());
-        for (JsonNode node : root) {
-          String name = node.path("name").asText("");
-          if (name.startsWith(prefix)) {
-            result.add(name);
-          }
-        }
-      } else {
-        throw new IllegalStateException(
-            "RabbitMQ management API returned status "
-                + response.statusCode()
-                + " for "
-                + resource);
-      }
-    } catch (IllegalStateException e) {
-      throw e;
-    } catch (Exception e) {
-      throw new IllegalStateException(
-          "Could not list " + resource + " from RabbitMQ management API: " + e.getMessage(), e);
-    }
-    return result;
+    return managementClient.listExchangeNamesWithPrefix(prefix);
   }
 
   // -- INTERNAL --
